pyproject.toml
[project]
name = "ipynb-translator"
version = "1.0.0"
description = "Jupyter Notebook translator using AWS Bedrock"
authors = [
    {name = "Daekeun Kim", email = "daekeun@amazon.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.40.21",
    "python-dotenv>=1.1.1",
    "nbformat>=5.10.4",
    "click>=8.2.1",
    "requests>=2.32.5",
    "fastmcp>=2.11.3",
]


[build-system]
requires = ["setuptools>=80.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["ipynb_translator*"]

[project.scripts]
ipynb-translate = "ipynb_translator.main:cli"
